# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Hotel AI OS Core API
  version: 0.2.3-pilot.6
  description: >-
    PRODUCT-V1.2 / TECH-V0.2 Sprint 2.1 UAT-candidate contract. Production requests use a
    trusted JWT/SSO bearer token; roles, permissions, organization scope and
    active position assignments are resolved by the server. X-Tenant-Id and
    X-Actor-Id are accepted only when DEV_HEADER_AUTH_ENABLED=true for local
    integration. Client-supplied role or organization-scope headers are never
    trusted.
servers:
  - url: http://localhost:8080
security:
  - BearerAuth: []
  - TenantHeader: []
    ActorHeader: []
...
paths:
...
  /api/v1/standards:
    get:
      tags: [Standard Center]
      operationId: listStandards
      summary: List standards and latest visible versions
      responses:
        '200': { $ref: '#/components/responses/ListResponse' }
    post:
      tags: [Standard Center]
      operationId: createStandard
      summary: Create a standard definition
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateStandard' }
      responses:
        '201': { $ref: '#/components/responses/ObjectResponse' }
  /api/v1/standards/{standardId}/versions:
    post:
      tags: [Standard Center]
      operationId: createStandardVersion
      summary: Create a draft standard version and applicability scopes
      parameters:
        - $ref: '#/components/parameters/StandardId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateStandardVersion' }
      responses:
        '201': { $ref: '#/components/responses/ObjectResponse' }
  /api/v1/standards/{standardId}/versions/{versionId}/publish:
    post:
      tags: [Standard Center]
      operationId: publishStandardVersion
      summary: Publish an immutable standard version
      parameters:
        - $ref: '#/components/parameters/StandardId'
        - $ref: '#/components/parameters/VersionId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PublishEffectiveWindow' }
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /api/v1/rules:
    get:
      tags: [Rule Engine]
      operationId: listRules
      summary: List enterprise rule definitions
      responses:
        '200': { $ref: '#/components/responses/ListResponse' }
    post:
      tags: [Rule Engine]
      operationId: createRule
      summary: Create an event-driven rule definition
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRule' }
      responses:
        '201': { $ref: '#/components/responses/ObjectResponse' }
  /api/v1/rules/{ruleId}:
    get:
      tags: [Rule Engine]
      operationId: getRule
      summary: Get a rule and versions
      parameters:
        - $ref: '#/components/parameters/RuleId'
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
  /api/v1/rules/{ruleId}/versions:
    post:
      tags: [Rule Engine]
      operationId: createRuleVersion
      summary: Create a draft rule version
      parameters:
        - $ref: '#/components/parameters/RuleId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRuleVersion' }
      responses:
        '201': { $ref: '#/components/responses/ObjectResponse' }
  /api/v1/rules/{ruleId}/versions/{versionId}:
    put:
      tags: [Rule Engine]
      operationId: updateRuleVersion
      summary: Update a draft rule version using expectedVersion
      parameters:
        - $ref: '#/components/parameters/RuleId'
        - $ref: '#/components/parameters/VersionId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateRuleVersion' }
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/rules/{ruleId}/versions/{versionId}/validate:
    post:
      tags: [Rule Engine]
      operationId: validateRuleVersion
      summary: Validate deterministic condition and action configuration
      parameters:
        - $ref: '#/components/parameters/RuleId'
        - $ref: '#/components/parameters/VersionId'
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
  /api/v1/rules/{ruleId}/versions/{versionId}/simulate:
    post:
      tags: [Rule Engine]
      operationId: simulateRuleVersion
      summary: Simulate a draft or published rule against supplied facts
      parameters:
        - $ref: '#/components/parameters/RuleId'
        - $ref: '#/components/parameters/VersionId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RuleSimulation' }
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
  /api/v1/rules/{ruleId}/versions/{versionId}/publish:
    post:
      tags: [Rule Engine]
      operationId: publishRuleVersion
      summary: Publish a validated rule version using expectedVersion
      parameters:
        - $ref: '#/components/parameters/RuleId'
        - $ref: '#/components/parameters/VersionId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PublishRuleVersion' }
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/rules/{ruleId}/versions/{versionId}/disable:
    post:
      tags: [Rule Engine]
      operationId: disableRuleVersion
      summary: Disable a rule version using expectedVersion
      parameters:
        - $ref: '#/components/parameters/RuleId'
        - $ref: '#/components/parameters/VersionId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DisableRuleVersion' }
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /api/v1/tasks:
    get:
      tags: [Task Center]
      operationId: listTasks
      summary: List tasks in effective data scope
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: orgUnitId, in: query, schema: { type: string, format: uuid } }
      responses:
        '200': { $ref: '#/components/responses/ListResponse' }
    post:
      tags: [Task Center]
      operationId: createTask
      summary: Create a task idempotently
      description: Assignee and reviewer must resolve to different active position assignments in scope.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTask' }
      responses:
        '201': { $ref: '#/components/responses/ObjectResponse' }
        '400': { $ref: '#/components/responses/Problem' }
  /api/v1/tasks/{taskId}:
    get:
      tags: [Task Center]
      operationId: getTask
      summary: Get task details and exact assignment participants
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
  /api/v1/tasks/{taskId}/timeline:
    get:
      tags: [Task Center]
      operationId: getTaskTimeline
      summary: Get an append-only task transition timeline
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200': { $ref: '#/components/responses/ListResponse' }
  /api/v1/tasks/{taskId}/evidence:
    post:
      tags: [Task Center]
      operationId: addTaskEvidence
      summary: Add task execution evidence metadata
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddTaskEvidence' }
      responses:
        '201': { $ref: '#/components/responses/ObjectResponse' }
  /api/v1/tasks/{taskId}/evidence/upload:
    post:
      tags: [Task Center]
      operationId: uploadTaskEvidence
      summary: Upload execution evidence for the exact assignee assignment
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - name: submittedByAssignmentId
          in: query
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '201': { $ref: '#/components/responses/ObjectResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/tasks/{taskId}/evidence/{evidenceId}/content:
    get:
      tags: [Task Center]
      operationId: downloadTaskEvidence
      summary: Download visible task evidence
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/EvidenceId'
      responses:
        '200':
          description: Evidence content
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/tasks/{taskId}/evidence/{evidenceId}:
    delete:
      tags: [Task Center]
      operationId: deleteTaskEvidence
      summary: Delete pre-submission evidence owned by the exact assignee assignment
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/EvidenceId'
        - name: actorAssignmentId
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/tasks/{taskId}/actions/{command}:
    post:
      tags: [Task Center]
      operationId: transitionTask
      summary: Apply an idempotent task state command using expectedVersion
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - name: command
          in: path
          required: true
          schema:
            type: string
            enum: [dispatch, acknowledge, start, submit-result, approve, rework, cancel]
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskCommand' }
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '400': { $ref: '#/components/responses/Problem' }
  /api/v1/tasks/sla/process:
    post:
      tags: [Task Center]
      operationId: processTaskSla
      summary: Mark overdue tasks and execute due escalation actions
      description: >-
        Requires task.dispatch. Marks active due tasks OVERDUE, records timeline entries,
        notifies reviewers, cancels scheduled escalations for terminal tasks, and executes
        due escalation rows with assignment-level notifications. Repeated processing does
        not repeat already executed or cancelled escalation rows.
      responses:
        '200':
          description: Counts produced by this processing pass
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskSlaProcessResult' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Trusted production JWT or SSO access token.
    TenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-Id
      description: Local integration only; requires DEV_HEADER_AUTH_ENABLED=true.
    ActorHeader:
      type: apiKey
      in: header
      name: X-Actor-Id
      description: Local integration only; requires DEV_HEADER_AUTH_ENABLED=true.
  parameters:
...
    StandardId: { name: standardId, in: path, required: true, schema: { type: string, format: uuid } }
...
    EvidenceId: { name: evidenceId, in: path, required: true, schema: { type: string, format: uuid } }
    VersionId: { name: versionId, in: path, required: true, schema: { type: string, format: uuid } }
...
    RuleId: { name: ruleId, in: path, required: true, schema: { type: string, format: uuid } }
...
    TaskId: { name: taskId, in: path, required: true, schema: { type: string, format: uuid } }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client-generated key; reuse returns the original semantic result.
      schema: { type: string, minLength: 1, maxLength: 255 }
  responses:
    ListResponse:
      description: List response
      content:
        application/json:
          schema:
            type: array
            items: { type: object, additionalProperties: true }
    ObjectResponse:
      description: Object response
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
    Problem:
      description: RFC 9457 problem detail
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Forbidden:
      description: Permission denied or organization scope violation
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: Optimistic version or idempotency conflict
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      properties:
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
...
    CreateStandard:
      type: object
      required: [categoryId, code, name]
      properties:
        categoryId: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        ownerOrgUnitId: { type: [string, 'null'], format: uuid }
        description: { type: [string, 'null'] }
    ApplicabilityScope:
      type: object
      required: [scopeType]
      properties:
        scopeType: { type: string, enum: [TENANT, BRAND, ORG_UNIT, ORG_TREE, POSITION] }
        brandId: { type: [string, 'null'], format: uuid }
        orgUnitId: { type: [string, 'null'], format: uuid }
        positionId: { type: [string, 'null'], format: uuid }
    CreateStandardVersion:
      type: object
      required: [title, items, scopes]
      properties:
        title: { type: string }
        items: { type: array, items: { type: object, additionalProperties: true } }
        evidenceRequirements: { type: array, items: { type: object, additionalProperties: true } }
        scoringRules: { type: object, additionalProperties: true }
        scopes: { type: array, minItems: 1, items: { $ref: '#/components/schemas/ApplicabilityScope' } }
    PublishEffectiveWindow:
      type: object
      required: [effectiveFrom]
      properties:
        effectiveFrom: { type: string, format: date-time }
        effectiveTo: { type: [string, 'null'], format: date-time }
...
    CreateRule:
      type: object
      required: [code, name, eventType]
      properties:
        code: { type: string }
        name: { type: string }
        eventType: { type: string }
        ownerOrgUnitId: { type: [string, 'null'], format: uuid }
        description: { type: [string, 'null'] }
    RuleScope:
      type: object
      required: [scopeType]
      properties:
        scopeType: { type: string }
        brandId: { type: [string, 'null'], format: uuid }
        orgUnitId: { type: [string, 'null'], format: uuid }
        positionId: { type: [string, 'null'], format: uuid }
    CreateRuleVersion:
      type: object
      required: [conditionAst, actions, scopes]
      properties:
        conditionAst: { type: object, additionalProperties: true }
        actions: {}
        priority: { type: [integer, 'null'] }
        cooldownMinutes: { type: [integer, 'null'] }
        scopes: { type: array, minItems: 1, items: { $ref: '#/components/schemas/RuleScope' } }
    UpdateRuleVersion:
      allOf:
        - $ref: '#/components/schemas/CreateRuleVersion'
        - type: object
          required: [expectedVersion]
          properties:
            expectedVersion: { type: integer, format: int64, minimum: 0 }
    PublishRuleVersion:
      type: object
      required: [expectedVersion, effectiveFrom]
      properties:
        expectedVersion: { type: integer, format: int64, minimum: 0 }
        effectiveFrom: { type: string, format: date-time }
        effectiveTo: { type: [string, 'null'], format: date-time }
    DisableRuleVersion:
      type: object
      required: [expectedVersion]
      properties:
        expectedVersion: { type: integer, format: int64, minimum: 0 }
    RuleSimulation:
      type: object
      required: [facts]
      properties:
        facts: { type: object, additionalProperties: true }
    CreateTask:
      type: object
      required: [orgUnitId, assigneeAssignmentId, reviewerAssignmentId, title]
      properties:
        orgUnitId: { type: string, format: uuid }
        assigneeAssignmentId: { type: string, format: uuid }
        reviewerAssignmentId: { type: string, format: uuid }
        standardVersionId: { type: [string, 'null'], format: uuid }
        workRecordId: { type: [string, 'null'], format: uuid }
        title: { type: string }
        description: { type: [string, 'null'] }
        priority: { type: [string, 'null'], enum: [LOW, NORMAL, HIGH, URGENT, null] }
        dueAt: { type: [string, 'null'], format: date-time }
        sourceSnapshot: { type: [object, 'null'], additionalProperties: true }
    TaskCommand:
      type: object
      required: [expectedVersion]
      properties:
        expectedVersion: { type: integer, format: int64, minimum: 0 }
        actorAssignmentId: { type: [string, 'null'], format: uuid }
        payload: { type: [object, 'null'], additionalProperties: true }
    AddTaskEvidence:
      type: object
      required: [submittedByAssignmentId, evidenceType]
      properties:
        submittedByAssignmentId: { type: string, format: uuid }
        evidenceType: { type: string }
        objectKey: { type: [string, 'null'] }
        originalName: { type: [string, 'null'] }
        mediaType: { type: [string, 'null'] }
        sizeBytes: { type: [integer, 'null'], format: int64 }
        sha256: { type: [string, 'null'], pattern: '^[a-fA-F0-9]{64}$' }
        structuredResult: { type: [object, 'null'], additionalProperties: true }
...
    TaskSlaProcessResult:
      type: object
      required: [overdueTasks, escalations, cancelledEscalations, notifications]
      properties:
        overdueTasks:
          type: integer
          minimum: 0
          description: Tasks newly marked OVERDUE in this pass.
        escalations:
          type: integer
          minimum: 0
          description: Due escalation rows successfully changed from SCHEDULED to EXECUTED.
        cancelledEscalations:
          type: integer
          minimum: 0
          description: Scheduled escalation rows cancelled because their tasks were terminal.
        notifications:
          type: integer
          minimum: 0
          description: Reviewer overdue and resolved-assignment escalation notifications created.
...